+++ source/crc32b.hpp
@@ -2,6 +2,6 @@
 #define CRC32B_CUSTOM_H_
 
 unsigned int crc32b(const char *message);
-float hash(const char *str);
+double hash(const char *str);
 
 #endif
